+++ src/main/java/sm/deeplink/dao/UserDao.java
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserDao extends JpaRepository<User, Integer> {
+public interface UserDao extends JpaRepository<User, Long> {
 
     User findByUsername(String username);
     User save(User user);
